+++ src/main/resources/db/migration/V2__add_multi_user_support.sql
@@ -6,7 +6,7 @@
 -- Table `users` (ユーザーマスタ)
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `users` (
-  `user_id` INT NOT NULL AUTO_INCREMENT,
+  `user_id` BIGINT NOT NULL AUTO_INCREMENT,
   `username` VARCHAR(255) NOT NULL UNIQUE,
   `password` VARCHAR(255) NOT NULL,
   PRIMARY KEY (`user_id`))
@@ -17,7 +17,7 @@ ENGINE = InnoDB;
 -- Add user_id to existing tables
 -- -----------------------------------------------------
 ALTER TABLE `accounts`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `account_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `account_id`,
   ADD INDEX `fk_accounts_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_accounts_users`
     FOREIGN KEY (`user_id`)
@@ -26,7 +26,7 @@ ALTER TABLE `accounts`
     ON UPDATE CASCADE;
 
 ALTER TABLE `categories`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `category_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `category_id`,
   ADD INDEX `fk_categories_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_categories_users`
     FOREIGN KEY (`user_id`)
@@ -35,7 +35,7 @@ ALTER TABLE `categories`
     ON UPDATE CASCADE;
 
 ALTER TABLE `transactions`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `transaction_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `transaction_id`,
   ADD INDEX `fk_transactions_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_transactions_users`
     FOREIGN KEY (`user_id`)
@@ -44,7 +44,7 @@ ALTER TABLE `transactions`
     ON UPDATE CASCADE;
 
 ALTER TABLE `recurring_transactions`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `recurring_transaction_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `recurring_transaction_id`,
   ADD INDEX `fk_recurring_transactions_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_recurring_transactions_users`
     FOREIGN KEY (`user_id`)
@@ -53,7 +53,7 @@ ALTER TABLE `recurring_transactions`
     ON UPDATE CASCADE;
 
 ALTER TABLE `borrows`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `borrow_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `borrow_id`,
   ADD INDEX `fk_borrows_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_borrows_users`
     FOREIGN KEY (`user_id`)
@@ -62,7 +62,7 @@ ALTER TABLE `borrows`
     ON UPDATE CASCADE;
 
 ALTER TABLE `goals`
-  ADD COLUMN `user_id` INT NOT NULL AFTER `goal_id`,
+  ADD COLUMN `user_id` BIGINT NOT NULL AFTER `goal_id`,
   ADD INDEX `fk_goals_users_idx` (`user_id` ASC) VISIBLE,
   ADD CONSTRAINT `fk_goals_users`
     FOREIGN KEY (`user_id`)
